+++ challenges/02-bank/submissions/helenna-massarra/src/IDPBank.java
@@ -1,168 +1,6 @@
 import java.util.Scanner;
 
-abstract class Account {
-    private static int accountCounter = 1000;
-    private int id;
-    protected double balance;
-
-    public Account() {
-        this.id = ++accountCounter;
-        this.balance = 0.0;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public double getBalance() {
-        return balance;
-    }
-
-    public abstract void deposit(double amount);
-
-    public abstract boolean withdraw(double amount);
-
-    public abstract void applyMonthlyProcessing();
-
-    @Override
-    public String toString() {
-        return "Número da Conta: " + id + "\n" +
-               "Saldo: R$ " + String.format("%.2f", balance);
-    }
-}
-
-class CheckingAccount extends Account {
-    private static final double MAINTENANCE_FEE = 10.0;
-    private static final double CREDIT_LIMIT = -500.0;
-
-    @Override
-    public void deposit(double amount) {
-        if (amount > 0) {
-            balance += amount;
-        }
-    }
-
-    @Override
-    public boolean withdraw(double amount) {
-        if (amount > 0 && (balance - amount) >= CREDIT_LIMIT) {
-            balance -= amount;
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public void applyMonthlyProcessing() {
-        balance -= MAINTENANCE_FEE;
-    }
-}
-
-class SavingsAccount extends Account {
-    private static final double INTEREST_RATE = 0.005;
-
-    @Override
-    public void deposit(double amount) {
-        if (amount > 0) {
-            balance += amount;
-        }
-    }
-
-    @Override
-    public boolean withdraw(double amount) {
-        if (amount > 0 && balance >= amount) {
-            balance -= amount;
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public void applyMonthlyProcessing() {
-        balance += balance * INTEREST_RATE;
-    }
-}
-
-class SalaryAccount extends Account {
-    @Override
-    public void deposit(double amount) {
-        System.out.println("Depósitos manuais não são permitidos em contas salário.");
-    }
-
-    @Override
-    public boolean withdraw(double amount) {
-        if (amount > 0 && balance >= amount) {
-            balance -= amount;
-            return true;
-        }
-        return false;
-    }
-
-    @Override
-    public void applyMonthlyProcessing() {
-        // Não há processamento mensal para contas salário
-    }
-}
-
-class Customer {
-    private String firstName;
-    private String lastName;
-    private String cpf;
-    private CheckingAccount checkingAccount;
-    private SavingsAccount savingsAccount;
-    private SalaryAccount salaryAccount;
-
-    public Customer(String firstName, String lastName, String cpf) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.cpf = cpf;
-    }
-
-    public void addAccount(Account account) {
-        if (account instanceof CheckingAccount) {
-            this.checkingAccount = (CheckingAccount) account;
-        } else if (account instanceof SavingsAccount) {
-            this.savingsAccount = (SavingsAccount) account;
-        } else if (account instanceof SalaryAccount) {
-            this.salaryAccount = (SalaryAccount) account;
-        }
-    }
-
-    public String getFullName() {
-        return firstName + " " + lastName;
-    }
-
-    public String displayInformation() {
-        StringBuilder info = new StringBuilder("\nNome: " + firstName + " " + lastName + "\n" +
-                                               "CPF: " + cpf + "\n");
-
-        if (checkingAccount != null) {
-            info.append("Conta Corrente:\n").append(checkingAccount).append("\n");
-        }
-        if (savingsAccount != null) {
-            info.append("Conta Poupança:\n").append(savingsAccount).append("\n");
-        }
-        if (salaryAccount != null) {
-            info.append("Conta Salário:\n").append(salaryAccount).append("\n");
-        }
-
-        return info.toString();
-    }
-
-    public Account getAccount(int accountType) {
-        switch (accountType) {
-            case 1:
-                return checkingAccount;
-            case 2:
-                return savingsAccount;
-            case 3:
-                return salaryAccount;
-            default:
-                return null;
-        }
-    }
-}
-
-class IDPBank {
+public class IDPBank {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         Customer customer = null;
